# docs/swagger.yaml
basePath: /api
definitions:
  handlers.loginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  handlers.tokenResponse:
    properties:
      refresh:
        type: string
      token:
        type: string
    type: object
  models.Character:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      game:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: Dokumentasi API untuk sistem karakter game
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://example.com/terms/
  title: Game Characters REST API
  version: "1.0"
paths:
  /characters:
    get:
      description: Mendapatkan list semua karakter dari database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Character'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Ambil semua karakter game
      tags:
      - characters
    post:
      consumes:
      - application/json
      parameters:
      - description: Character Data
        in: body
        name: character
        required: true
        schema:
          $ref: '#/definitions/models.Character'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Character'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Tambah karakter baru
      tags:
      - characters
  /characters/{id}:
    delete:
      parameters:
      - description: Character ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Hapus karakter
      tags:
      - characters
    get:
      parameters:
      - description: Character ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Character'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Ambil karakter berdasarkan ID
      tags:
      - characters
    put:
      consumes:
      - application/json
      parameters:
      - description: Character ID
        in: path
        name: id
        required: true
        type: integer
      - description: Character Data
        in: body
        name: character
        required: true
        schema:
          $ref: '#/definitions/models.Character'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Character'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update karakter
      tags:
      - characters
  /login:
    post:
      consumes:
      - application/json
      description: Login menggunakan username & password, menghasilkan access token
        + refresh token
      parameters:
      - description: Username dan Password
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/handlers.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.tokenResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login
      tags:
      - auth
  /logout:
    post:
      description: Logout user, menghapus access token & refresh token (client-side
        dan cookie akan dihapus)
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"